+++ Net/Connection/SocketConnection.cpp
@@ -133,8 +133,8 @@ int SocketConnection::asynRead(int theFd, short theEvt)
 void SocketConnection::onRead(int theFd, short theEvt)
 {
     char buffer[1024]= {0};
-    size_t readBufferLeft = inputQueueM.unusedSize();
-    size_t readLen = (readBufferLeft < sizeof(buffer)) ? readBufferLeft : sizeof(buffer);
+    unsigned readBufferLeft = inputQueueM.unusedSize();
+    unsigned readLen = (readBufferLeft < sizeof(buffer)) ? readBufferLeft : sizeof(buffer);
     if (readLen == 0)
     {
         if (!stopReadingM)
@@ -159,8 +159,8 @@ void SocketConnection::onRead(int theFd, short theEvt)
         _close();
         return;
     }
-    size_t putLen = inputQueueM.put(buffer, len);
-    assert(putLen == (size_t)len);
+    unsigned putLen = inputQueueM.put(buffer, len);
+    assert(putLen == (unsigned)len);
 
     while(Net::Buffer::BufferOkE == inputQueueM.getStatus()
         || Net::Buffer::BufferLowE == inputQueueM.getStatus() )
@@ -173,7 +173,7 @@ void SocketConnection::onRead(int theFd, short theEvt)
             break;
         }
         putLen = inputQueueM.put(buffer, len);
-        assert(putLen == (size_t)len);
+        assert(putLen == (unsigned)len);
     }
 
     if (Net::Buffer::BufferHighE == inputQueueM.getStatus()
@@ -191,12 +191,12 @@ void SocketConnection::onRead(int theFd, short theEvt)
 }
 //-----------------------------------------------------------------------------
 
-size_t SocketConnection::getInput(char* const theBuffer, const size_t theLen)
+unsigned SocketConnection::getInput(char* const theBuffer, const unsigned theLen)
 {
     if (CloseE == statusM)
         return 0;
 
-    size_t len = inputQueueM.get(theBuffer, theLen);
+    unsigned len = inputQueueM.get(theBuffer, theLen);
     if (stopReadingM)
     {
         Net::Buffer::BufferStatus postBufferStatus = inputQueueM.getStatus();
@@ -214,12 +214,12 @@ size_t SocketConnection::getInput(char* const theBuffer, const size_t theLen)
 
 //-----------------------------------------------------------------------------
 
-size_t SocketConnection::getnInput(char* const theBuffer, const size_t theLen)
+unsigned SocketConnection::getnInput(char* const theBuffer, const unsigned theLen)
 {
     if (CloseE == statusM)
         return 0;
 
-    size_t len = inputQueueM.getn(theBuffer, theLen);
+    unsigned len = inputQueueM.getn(theBuffer, theLen);
     if (stopReadingM)
     {
         Net::Buffer::BufferStatus postBufferStatus = inputQueueM.getStatus();
@@ -237,7 +237,7 @@ size_t SocketConnection::getnInput(char* const theBuffer, const size_t theLen)
 
 //-----------------------------------------------------------------------------
 
-size_t SocketConnection::peeknInput(char* const theBuffer, const size_t theLen)
+unsigned SocketConnection::peeknInput(char* const theBuffer, const unsigned theLen)
 {
     if (CloseE == statusM)
         return 0;
@@ -247,7 +247,7 @@ size_t SocketConnection::peeknInput(char* const theBuffer, const size_t theLen)
 
 //-----------------------------------------------------------------------------
 
-size_t SocketConnection::sendn(char* const theBuffer, const size_t theLen)
+unsigned SocketConnection::sendn(char* const theBuffer, const unsigned theLen)
 {
     if (CloseE == statusM)
         return 0;
@@ -255,7 +255,7 @@ size_t SocketConnection::sendn(char* const theBuffer, const size_t theLen)
     if (0 == theLen || NULL == theBuffer)
         return 0;
 
-    size_t len = 0;
+    unsigned len = 0;
     {
         boost::lock_guard<boost::mutex> lock(outputQueueMutexM);
         len = outputQueueM.putn(theBuffer, theLen);
@@ -303,7 +303,7 @@ void SocketConnection::onWrite(int theFd, short theEvt)
         }
     }
     char buffer[1024]= {0};
-    size_t peekLen = outputQueueM.peek(buffer, sizeof(buffer));
+    unsigned peekLen = outputQueueM.peek(buffer, sizeof(buffer));
     int writeLen = 0;
     while (CloseE != statusM && peekLen > 0)
     {
